# pops-project-manager-api/contrato/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "POPS Project Manager API",
        "description": "API para gerenciamento de projetos e skills",
        "contact": {
            "name": "POPS Team",
            "email": "gyulia.piqueira@sptech.school"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8082/api",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/skills/enable/{id}": {
            "put": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "enableSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/skills/disable/{id}": {
            "put": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "disableSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/skill-types/enable/{id}": {
            "put": {
                "tags": [
                    "skill-type-controller"
                ],
                "operationId": "enableType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/skill-types/disable/{id}": {
            "put": {
                "tags": [
                    "skill-type-controller"
                ],
                "operationId": "disableType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/project-types/enable/{id}": {
            "put": {
                "tags": [
                    "project-type-controller"
                ],
                "operationId": "enableType_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/project-types/disable/{id}": {
            "put": {
                "tags": [
                    "project-type-controller"
                ],
                "operationId": "disableType_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/skill-types": {
            "get": {
                "tags": [
                    "skill-type-controller"
                ],
                "operationId": "listActiveTypes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "skill-type-controller"
                ],
                "operationId": "createType",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SkillTypeCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SkillTypeEntity"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/project-types": {
            "get": {
                "tags": [
                    "project-type-controller"
                ],
                "operationId": "listActiveTypes_1",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "project-type-controller"
                ],
                "operationId": "createType_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectTypeCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectTypeEntity"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/skills/{id}": {
            "get": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "getSkillById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SkillResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "deleteSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "updateSkill",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SkillUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/skill-types/{id}": {
            "get": {
                "tags": [
                    "skill-type-controller"
                ],
                "operationId": "getTypeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SkillTypeResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "skill-type-controller"
                ],
                "operationId": "deleteType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "skill-type-controller"
                ],
                "operationId": "updateType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SkillTypeUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/project-types/{id}": {
            "get": {
                "tags": [
                    "project-type-controller"
                ],
                "operationId": "getTypeById_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectTypeResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "project-type-controller"
                ],
                "operationId": "deleteType_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "project-type-controller"
                ],
                "operationId": "updateType_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectTypeUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/skills/type/{typeId}": {
            "get": {
                "tags": [
                    "skill-controller"
                ],
                "operationId": "listSkillsByType",
                "parameters": [
                    {
                        "name": "typeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SkillTypeEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name"
                ]
            },
            "SkillTypeCreateRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    }
                },
                "required": [
                    "name"
                ]
            },
...
            "ProjectTypeEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name"
                ]
            },
            "ProjectTypeCreateRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    }
                },
                "required": [
                    "name"
                ]
            },
...
            "SkillUpdateRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "typeId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "SkillTypeUpdateRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    }
                }
            },
...
            "ProjectTypeUpdateRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    }
                }
            },
...
            "SkillResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/SkillTypeResponse"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name",
                    "type"
                ]
            },
            "SkillTypeResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name"
                ]
            },
...
            "ProjectTypeResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "active": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "active",
                    "name"
                ]
            },
...
        }
    }
}